Changes according to the comments about CACTO Refactoring v3

+++ utils.py
@@ -9,7 +9,7 @@ def de_normalize_tensor(state, state_norm_arr):
     ''' Retrieve state from normalized state - tensor '''
     state_time = tf.concat([tf.zeros([state.shape[0], state.shape[1]-1]), tf.reshape((state[:,-1]+1)*state_norm_arr[-1]/2,[state.shape[0],1])],1)
     state_no_time = state * state_norm_arr
-    mask = tf.cast(tf.concat([tf.ones([state.shape[0], state.shape[1]-1]), tf.zeros([state.shape[0], 1])],1),tf.float32)
+    mask = tf.concat([tf.ones([state.shape[0], state.shape[1]-1]), tf.zeros([state.shape[0], 1])],1)
     state_not_norm = state_no_time * mask + state_time * (1 - mask)
 
     return state_not_norm
@@ -18,7 +18,7 @@ def normalize_tensor(state, state_norm_arr):
     ''' Retrieve state from normalized state - tensor '''
     state_norm_time = tf.concat([tf.zeros([state.shape[0], state.shape[1]-1]), tf.reshape(((state[:,-1]) / state_norm_arr[-1])*2 - 1,[state.shape[0],1])],1)
     state_norm_no_time = state / state_norm_arr
-    mask = tf.cast(tf.concat([tf.ones([state.shape[0], state.shape[1]-1]), tf.zeros([state.shape[0], 1])],1),tf.float32)
+    mask = tf.concat([tf.ones([state.shape[0], state.shape[1]-1]), tf.zeros([state.shape[0], 1])],1)
     state_norm = state_norm_no_time * mask + state_norm_time * (1 - mask)
 
     return state_norm
